+++ docs/source/changelog.rst
@@ -14,6 +14,7 @@ Version (dev)
 * Add support for saving/loading VTF hotspot data.
 * Don't backslash-escape single quote characters.
 * Consult the various app IDs when looking for folders for searchpaths.
+* Deprecate :meth:`UVAxis.rotate()`.
 
 -------------
 Version 2.6.1